requirements.txt
Flask
flask-restx
flask-sqlalchemy
flask-jwt-extended
flask-migrate
psycopg2-binary
python-dotenv
flask-cors